✨ Add TopBar UI component with menu functionality

+++ src/editor/main.cpp
@@ -1,6 +1,7 @@
 #include <vector>
 
 #include "graph.hpp"
+#include "ui/top_bar.hpp"
 
 int main() {
     InitWindow(800, 600, "Nebula Test");
@@ -8,6 +9,50 @@ int main() {
 
     // Create graph
     core::Graph graph;
+
+    std::vector<editor_ui::Menu> menus;
+
+    menus.push_back(editor_ui::Menu{
+        "File",
+        {
+            editor_ui::MenuItem{"New", [&graph]() { graph = core::Graph{}; }},
+            editor_ui::MenuItem{"Save", []() {
+                TraceLog(LOG_INFO, "Save requested from top bar");
+            }},
+            editor_ui::MenuItem{"Load", []() {
+                TraceLog(LOG_INFO, "Load requested from top bar");
+            }},
+            editor_ui::MenuItem{"Quit", []() { CloseWindow(); }},
+        },
+    });
+
+    menus.push_back(editor_ui::Menu{
+        "Add",
+        {
+            editor_ui::MenuItem{"Variable", [&graph, &cursor_postion]() {
+                graph.AddNode(core::NodeBase::NodeKind::kVariable,
+                              {cursor_postion.x - 50, cursor_postion.y - 25});
+            }},
+            editor_ui::MenuItem{"Literal", [&graph, &cursor_postion]() {
+                graph.AddNode(core::NodeBase::NodeKind::kLiteral,
+                              {cursor_postion.x - 50, cursor_postion.y - 25});
+            }},
+        },
+    });
+
+    menus.push_back(editor_ui::Menu{
+        "Edit",
+        {
+            editor_ui::MenuItem{"Duplicate selection", [&graph]() {
+                graph.DuplicateSelectedNode();
+            }},
+            editor_ui::MenuItem{"Delete under cursor", [&graph]() {
+                graph.DeleteWithMouse();
+            }},
+        },
+    });
+
+    editor_ui::TopBar top_bar(std::move(menus));
 
     // Create a visual node
     auto node1 = graph.AddNode(core::NodeBase::NodeKind::kLiteral, {500, 100});
@@ -43,19 +88,28 @@ int main() {
         }
 
         cursor_postion = GetMousePosition();
+        top_bar.Update();
 
-        graph.DeleteWithMouse();
-        graph.CheckNodeMovement();
+        if (!top_bar.BlocksGraphInput()) {
+            graph.DeleteWithMouse();
+            graph.CheckNodeMovement();
+        }
 
         BeginDrawing();
         ClearBackground(RAYWHITE);
 
         // Draw visual nodes in the graph
-        graph.SelectWithMouse();
+        if (!top_bar.BlocksGraphInput()) {
+            graph.SelectWithMouse();
+        }
         graph.Draw();
 
-        graph.LinkingWithMouse();
-        graph.HandleContextMenu();
+        if (!top_bar.BlocksGraphInput()) {
+            graph.LinkingWithMouse();
+            graph.HandleContextMenu();
+        }
+
+        top_bar.Draw();
 
         EndDrawing();
     }
